+++ protocol-code-generator/src/main/java/dev/cirras/xml/ProtocolEnum.java
@@ -1,6 +1,5 @@
 package dev.cirras.xml;
 
-import jakarta.xml.bind.Unmarshaller;
 import jakarta.xml.bind.annotation.XmlAttribute;
 import jakarta.xml.bind.annotation.XmlElement;
 import jakarta.xml.bind.annotation.XmlRootElement;
@@ -16,8 +15,6 @@ public final class ProtocolEnum {
   @XmlAttribute(required = true)
   private String type;
 
-  @XmlAttribute private boolean clamp;
-
   @XmlElement(name = "value", required = true)
   private final List<ProtocolValue> values = new ArrayList<>();
 
@@ -31,22 +28,11 @@ public final class ProtocolEnum {
     return type;
   }
 
-  public boolean isClamp() {
-    return clamp;
-  }
-
   public List<ProtocolValue> getValues() {
     return values;
   }
 
   public Optional<ProtocolComment> getComment() {
     return Optional.ofNullable(comment);
-  }
-
-  @SuppressWarnings("unused")
-  private void afterUnmarshal(Unmarshaller unmarshaller, Object parent) {
-    if (values.stream().filter(ProtocolValue::isDefault).count() > 1) {
-      throw new ProtocolXmlError("Only one default enum value is allowed.");
-    }
   }
 }